move start functions back to todo_app
