+++ prod-start.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-poetry run gunicorn wsgi:start -b 0.0.0.0:$PORT
+poetry run gunicorn todo_app.wsgi:start -b 0.0.0.0:$PORT
